fix(check): probe the active embedding profile, not only the genesis seed
- tribal check now probes the active profile's endpoint (the live
  read/write identity, §5.7) when a corpus exists, falling back to the
  genesis seed only when there is no active profile, and the
  genesis-vs-active divergence note compares base_url and dimensions as
  well as provider and model (B2).
- The reindex check comment is reconciled from "stuck" to "a live
  reindex run", matching the shipped any-live-run warn condition (C8).
- The cumulative reindex_runs tallies (items/tags embedded and
  quarantined) widen to BIGINT to stay clear of the INT ceiling at scale;
  the single-snapshot *_enumerated columns stay INT (C6).
- Removed the pre-existing em dashes in scope.rs doc comments (C9).

+++ crates/tribal-db/migrations/20260214164025_initial_schema.sql
@@ -340,12 +340,18 @@ CREATE TABLE reindex_runs (
                                    CHECK (state IN ('queued', 'running', 'completed',
                                                     'superseded', 'aborted', 'failed')),
     initiated_by_principal_id  UUID NOT NULL REFERENCES principals(id),
+    -- The *_enumerated columns are single backlog snapshots, so INT suffices.
+    -- The *_embedded and *_quarantined columns are cumulative tallies summed
+    -- across every catch-up pass, so they can exceed the corpus size and must be
+    -- BIGINT to stay clear of the INT ceiling at scale. Editing the initial
+    -- schema is the branch convention: databases are recreated, there are no
+    -- users.
     items_enumerated           INT,
-    items_embedded             INT NOT NULL DEFAULT 0,
+    items_embedded             BIGINT NOT NULL DEFAULT 0,
     tags_enumerated            INT,
-    tags_embedded              INT NOT NULL DEFAULT 0,
-    items_quarantined          INT NOT NULL DEFAULT 0,
-    tags_quarantined           INT NOT NULL DEFAULT 0,
+    tags_embedded              BIGINT NOT NULL DEFAULT 0,
+    items_quarantined          BIGINT NOT NULL DEFAULT 0,
+    tags_quarantined           BIGINT NOT NULL DEFAULT 0,
     error_message              TEXT,
     trace_context              TEXT CHECK (trace_context IS NULL OR char_length(trace_context) <= 128),
     started_at                 TIMESTAMPTZ NOT NULL DEFAULT now(),
